Initial hacker documentation
This was such a good resource in Docker that it would be a shame to
lose it. Unfortunately, we can't just link to the corresponding
information in the Docker project as a number of aspects are bound to
be different here.

Docker-DCO-1.1-Signed-off-by: Glyn Normington <gnormington@gopivotal.com> (github: glyn)