Initial integration of libct into libcontainer
This patch adds new factory and container classes to create and manage
containers via LibCT.

This implementation is incomplete, but I think it's enough to start an
integration process.

LibCT is a library for creating and managing containers. It has ability
to add different backends. Currently it has only a native backend and a
backend for OpenVZ and PCS.

Cc: Michael Crosby <michael@docker.com>
Cc: Victor Marmol <vmarmol@google.com>
Cc: Pavel Emelyanov <xemul@parallels.com>
Signed-off-by: Andrey Vagin <avagin@openvz.org>